FSPT-1012: consistent URL and folder structure
Change `/grant` to `/grants` in the reports URL and also moves the
reports route handler into its own file as it could be considered a
separate part of the application.
