+++ app/access_grant_funding/routes/misc.py
@@ -11,14 +11,9 @@ from app.common.auth.decorators import (
     is_access_org_member,
 )
 from app.common.data import interfaces
-from app.common.data.interfaces.collections import (
-    get_all_submissions_with_mode_for_collection_with_full_schema,
-)
 from app.common.data.interfaces.grant_recipients import get_grant_recipient
-from app.common.data.interfaces.grants import get_grant
 from app.common.data.interfaces.organisations import get_organisation
-from app.common.data.types import RoleEnum, SubmissionModeEnum
-from app.common.helpers.collections import SubmissionHelper
+from app.common.data.types import RoleEnum
 
 
 @access_grant_funding_blueprint.route("/", methods=["GET"])
@@ -81,40 +76,7 @@ def list_organisations() -> ResponseReturnValue:
 
 
 @access_grant_funding_blueprint.route(
-    "/organisation/<uuid:organisation_id>/grant/<uuid:grant_id>/reports", methods=["GET"]
-)
-@has_access_grant_role(RoleEnum.MEMBER)
-def list_reports(organisation_id: UUID, grant_id: UUID) -> ResponseReturnValue:
-    grant = get_grant(grant_id=grant_id)
-
-    grant_recipient = get_grant_recipient(grant_id, organisation_id)
-
-    # TODO refactor when we persist the collection status and/or implement multiple rounds
-    submissions = []
-    for report in grant.access_reports:
-        submissions.extend(
-            [
-                SubmissionHelper(submission=submission)
-                for submission in get_all_submissions_with_mode_for_collection_with_full_schema(
-                    collection_id=report.id,
-                    submission_mode=SubmissionModeEnum.LIVE,
-                    grant_recipient_id=grant_recipient.id,
-                )
-            ]
-        )
-
-    return render_template(
-        "access_grant_funding/report_list.html",
-        reports=grant.access_reports,
-        organisation_id=organisation_id,
-        grant=grant,
-        submissions=submissions,
-        grant_recipient=grant_recipient,
-    )
-
-
-@access_grant_funding_blueprint.route(
-    "/organisation/<uuid:organisation_id>/grant/<uuid:grant_id>/users", methods=["GET"]
+    "/organisation/<uuid:organisation_id>/grants/<uuid:grant_id>/users", methods=["GET"]
 )
 @has_access_grant_role(RoleEnum.MEMBER)
 def list_grant_team(organisation_id: UUID, grant_id: UUID) -> ResponseReturnValue:
