+++ app/access_grant_funding/routes/reports.py
@@ -1,0 +1,42 @@
+from uuid import UUID
+
+from flask import render_template
+from flask.typing import ResponseReturnValue
+
+from app.access_grant_funding.routes import access_grant_funding_blueprint
+from app.common.auth.decorators import has_access_grant_role
+from app.common.data.interfaces.collections import get_all_submissions_with_mode_for_collection_with_full_schema
+from app.common.data.interfaces.grant_recipients import get_grant_recipient
+from app.common.data.types import RoleEnum, SubmissionModeEnum
+from app.common.helpers.collections import SubmissionHelper
+
+
+@access_grant_funding_blueprint.route(
+    "/organisation/<uuid:organisation_id>/grants/<uuid:grant_id>/reports", methods=["GET"]
+)
+@has_access_grant_role(RoleEnum.MEMBER)
+def list_reports(organisation_id: UUID, grant_id: UUID) -> ResponseReturnValue:
+    grant_recipient = get_grant_recipient(grant_id, organisation_id)
+
+    # TODO refactor when we persist the collection status and/or implement multiple rounds
+    submissions = []
+    for report in grant_recipient.grant.access_reports:
+        submissions.extend(
+            [
+                SubmissionHelper(submission=submission)
+                for submission in get_all_submissions_with_mode_for_collection_with_full_schema(
+                    collection_id=report.id,
+                    submission_mode=SubmissionModeEnum.LIVE,
+                    grant_recipient_id=grant_recipient.id,
+                )
+            ]
+        )
+
+    return render_template(
+        "access_grant_funding/report_list.html",
+        reports=grant_recipient.grant.access_reports,
+        organisation_id=organisation_id,
+        grant=grant_recipient.grant,
+        submissions=submissions,
+        grant_recipient=grant_recipient,
+    )
